requirements.txt
django>=4.0,<5.0
djangorestframework>=3.14,<4.0
psycopg2>=2.9,<3.0
django-cors-headers>=3.14,<4.0
python-dotenv>=0.19,<1.0
gunicorn